Tidy server tool registrar wiring
